move scrobblers into their own package
