+++ server/scrobble/scrobble.go
@@ -1,0 +1,9 @@
+package scrobble
+
+import (
+	"go.senan.xyz/gonic/server/db"
+)
+
+type Scrobbler interface {
+	Scrobble(user *db.User, track *db.Track, stampMili int, submission bool) error
+}
